+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 - #61 - Autodiscover Kafka brokers based on Zookeeper data
 - #63 - Graceful shutdown with current offset state during data processing
 - #65 - Example of NewRelic monitor is outdated
-- Gem dump
+- Gem dump x2
 - Rubocop remarks
 - worker_timeout config option has been removed. It now needs to be defined manually by the framework user because WorkerGlass::Timeout can be disabled and we cannot use Karafka settings on a class level to initialize user code stuff
 - #71 - Setup should be executed after user code is loaded
+- Moved setup logic under setup/Setup namespace
+- Better defaults handling
 
 ## 0.4.0
 - Added WaterDrop gem with default configuration
